+++ example/main.dart
@@ -1,46 +1,72 @@
 import 'package:diffutil_dart/diffutil.dart' as diffutil;
 
+/// Note: This is a dart-only project, running flutter pub get will fail with an error message of a missing
+/// pubspec.yaml in the example folder. Just run pub get instead.
+
 void main() {
   final oldList = [1, 2, 3];
   final newList = [2, 3, 4];
 
-  print("difference between $oldList and $newList, with move detection:");
+  print("difference between $oldList and $newList, without move detection:");
 
-  final listDiff = diffutil.calculateListDiff(oldList, newList);
+  final listDiff = diffutil.calculateListDiff(oldList, newList).getUpdates();
 
   // use the diff using a list of diff objects
-  for (final update in listDiff.getUpdates())
+  for (final update in listDiff)
     update.when(
-      insert: (pos, count) => print("inserted $count on $pos"),
-      remove: (pos, count) => print("removed $count on $pos"),
+      insert: (pos, count) => print("inserted $count item on $pos"),
+      remove: (pos, count) => print("removed $count item on $pos"),
       change: (pos, payload) => print("changed on $pos with payload $payload"),
-      move: (from, to) => print("move $from to $to"),
+      move: (from, to) => print("move from $from to $to"),
     );
 
-  // use the diff using a custom callback class
-
-  listDiff.dispatchUpdatesTo(MyListCallback());
+  print("changeset: $listDiff");
 
   final oldList2 = [1, 2, 3];
   final newList2 = [1, 3, 2];
 
   print("\n");
 
-  print("difference between $oldList and $newList, without move detection:");
+  print("difference between $oldList2 and $newList2, without move detection:");
 
-  final listDiff2 =
-      diffutil.calculateListDiff(oldList2, newList2, detectMoves: false);
+  final listDiff2 = diffutil
+      .calculateListDiff(oldList2, newList2, detectMoves: false)
+      .getUpdates();
 
-  listDiff2.dispatchUpdatesTo(MyListCallback());
+  print("changeset: $listDiff2");
 
   print("\n");
 
-  print("difference between $oldList and $newList, with move detection:");
+  print("difference between $oldList2 and $newList2, with move detection:");
 
   final listDiff3 =
-      diffutil.calculateListDiff(oldList2, newList2, detectMoves: true);
+      diffutil.calculateListDiff(oldList2, newList2, detectMoves: true).getUpdates();
 
-  listDiff3.dispatchUpdatesTo(MyListCallback());
+  print("changeset: $listDiff3");
+
+  print("\n");
+
+  final oldList3 = [];
+  final newList3 = [1,2,3];
+
+  print("difference between $oldList3 and $newList3, batched:");
+
+  final listDiff4 = diffutil.calculateListDiff(oldList3, newList3).getUpdates(batch: true);
+
+  print("changeset: $listDiff4");
+
+  print("\n");
+
+  print("difference between $oldList3 and $newList3, unbatched:");
+
+  final listDiff5 = diffutil.calculateListDiff(oldList3, newList3).getUpdates(batch: false);
+
+  print("changeset: $listDiff5");
+
+  print("\n");
+
+
+
 }
 
 class MyListCallback implements diffutil.ListUpdateCallback {