+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.0
+
+Major revamp to make the libary more Dart-y and less cumbersome to use!
+
+- Add ability to calculate the changeset as list of DiffUpdate object, where each object is of type Insert, Remove, Change or Insert
+- Deprecated the old way the get the changeset via ListUpdateCallback.
+- Ability to turn off changeset batching (See README for an explanation on changeset batching)
+- Improved the example
+
 ## 0.1.0+1
 
 - more tests